Remove UserFromAuthorizationToken; add legacy method for now
